feat: refactor commands
- move commands to a separate module
- move command handling to a separate file
- use a trait for enforcing similar command structure
- move around tests to where necessary
- only focus on handling errors in the main function
